Move userApi package to bot
