+++ bot/userApi/keyPair.go
